+++ src/markdown-parser.js
@@ -15,7 +15,8 @@ import {
   TableRow,
   TableCell,
   List,
-  ListItem,
+  OrderedListItem,
+  UnorderedListItem,
   Hr,
   Html,
   Blockquote,
@@ -38,7 +39,8 @@ export const defaultRenderers = {
   tablerow: TableRow,
   tablecell: TableCell,
   list: List,
-  listitem: ListItem,
+  orderedlistitem: OrderedListItem,
+  unorderedlistitem: UnorderedListItem,
   hr: Hr,
   html: Html,
   blockquote: Blockquote,
