+++ src/renderers/OrderedListItem.svelte
