+++ src/renderers/index.js
@@ -13,7 +13,8 @@ export { default as TableBody } from './TableBody.svelte'
 export { default as TableRow } from './TableRow.svelte'
 export { default as TableCell } from './TableCell.svelte'
 export { default as List } from './List.svelte'
-export { default as ListItem } from './ListItem.svelte'
+export { default as OrderedListItem } from './OrderedListItem.svelte'
+export { default as UnorderedListItem } from './UnorderedListItem.svelte'
 export { default as Hr } from './Hr.svelte'
 export { default as Html } from './Html.svelte'
 export { default as Blockquote } from './Blockquote.svelte'
